Add functions to GraphQL (#5279)
