+++ crates/core/src/gql/mod.rs
@@ -3,7 +3,9 @@
 pub mod cache;
 pub mod error;
 mod ext;
+mod functions;
 pub mod schema;
+mod tables;
 mod utils;
 
 pub use error::GqlError;
